[DEV] pomp_loop_remove: always remove fd from list
Always remove fd from the internal linked list when pomp_loop_remove is called.

Not removing fd from the list could be harmful in some circumstances:

Example:
If the fd has been closed before calling pomp_loop_remove,
epoll (linux implementation) will return a EBADF (to indicate the fd
is no more valid and has already been removed from epoll by the kernel).
In this case, it is an error to not remove it from the linked list,
because the user will not be able to add the same fd again.
A dangling fd has been created and stay in the pomp_loop.

Change-Id: I96404e9dc98a861aefb1356b7493d3ac345a50af
Reviewed-on: https://gerrit.parrot.biz/48394
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>