[DEBUG] Connection can be not referenced in .hpp when receiving udp message
In hpp API, we look for the connection associated with received message. But
with dgam connection type, this connection can be not present in connections
array.
Trying to access the value of end() iterator did a segfault.

This commit handles this case, sending a nullptr in case of not referenced
connection in hpp side.

Change-Id: Id228caf991113349fc6a6f04d321ba49555059a2
Reviewed-on: https://gerrit.parrot.biz/64183
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>