[DEV] loop: improve event processing speed by using a hash table of fd
On Linux, with a 1ms timer added to pomp loop, perf reports

   6,99%  patric    libpomp.so     [.] pomp_loop_epoll_do_wait_and_process

In the same program, without the timer, perf reports:

   0,53%  patric    libpomp.so     [.] pomp_loop_epoll_do_wait_and_process

Searching the pomp file descriptor associated to the file
descriptor for each event is a bit slow, being linear.

Especially as it's searching in a linked list, it involves
a lot of pointer chasing.

Replacing the linked list with a sorted array or a binary
tree would likely improve the performances but at the cost
of complexity.

Here, an adhoc fixed size hash table, using linked list to
handle collision, is introduced, allowing to minimize the
amount of changes to be made in pomp_loop by reusing the
linear search mechanism.

The hash table is dimensionned so that it can fit the file
descriptor added to the loop by programs running on target:
14.8 file descriptors are added on average, 90% of the programs
on the target are using 32.8 or less file descriptors, 95%
of them are using 48.7 or less, and 98% are using 94.2 or
less. The minimum being 2, maximum 113, and the median 7.

The hash table is made 128 slot/bucket to specifically to
allow the programs  using the most file descriptors (93, 113)
to be more efficient. It also reduces the likelyhood of
collision for the other programs.

So 128 entries should cover most needs, while
1) being a power of two.
2) making the table only 1KBytes;

Hashing the file descriptor is mostly required only on
Windows, where file descriptor are either
1) SOCKET, an UINT_PTR, with value multiple of 4 and in
   range [0..0x400000] (verified on Windows 10 and WINE)
2) pointer, with value multiple of 16 on x86_64 or 8 on
   x86 (verified on Windows 10 and WINE, also range is
    irrelevant because it depends on the address space
    of the process).

On Unix, the file descriptors are small integers, with
either a value increasing on each new allocation, or a
free, lower one being reused, ensuring the hash table
would be filled quite uniformally.

Anyway, hashing should improve spreading the file descriptor
in slots, in the case the file descriptor follow some
pattern, for example, being mostly even or mostly odd.

A very simple, basic hash function is used to hash the file
descriptor values. It's using constants randomly generated
until the hash function compute a minimum colliding indexes
for file descriptors in [0..127] range, then be optimal in
range [0..1023], [0..3ffffc] for multiple of 4, [0..2^31-1]
for multiple of 8 and multiple of 16.

Please note it's not a proper general purpose hash function,
first because it's designed for the length of the hash table,
second, because it lacks properties that would make its
output indistinguishable from random.

After the patch, perf reports

   2,04%  patric    libpomp.so     [.] pomp_loop_epoll_do_wait_and_process