[DEV] pomp_ctx: use file descriptor returned by event loop
Don't use ctx's file descriptor directly, use the one
reported by the loop ... but check it against the one
stored in the associated context. Also context type
must match to make the check.

It will prevent using -1 as file descriptor when context
is being shutdown unexpectedly.

But this is mostly designed to help diagnose corruptions
on contexts.